requirements-optional.txt
pysheds
richdem
gdal
geopandas
optuna>=4.0
asyncpg
psycopg2
psycopg2-binary
plotly
matplotlib
